Roll back bash version to v4.4.023-1
There appears to be an issue signing the 'sh.exe' binary in the newest
versions of Git for Windows. This was traced back to the most recent upgrade
(d436fd07c (Update 3 packages, 2022-01-18)) of the 'bash' package.

For now, roll back to the v4.4.023-1 version to allow signable Git for
Windows artifacts to be generated.

Signed-off-by: Victoria Dye <vdye@github.com>